refactor: catalina, coyote 성격에 맞게 패키지 이동
