+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,0 +1,9 @@
+package org.apache.catalina.controller;
+
+import org.apache.catalina.response.HttpResponse;
+import org.apache.catalina.request.HttpRequest;
+
+public interface Controller {
+
+    void service(HttpRequest request, HttpResponse response);
+}
