+++ tomcat/src/main/java/org/apache/catalina/controller/StaticController.java
@@ -1,0 +1,14 @@
+package org.apache.catalina.controller;
+
+import org.apache.catalina.http.StatusCode;
+import org.apache.catalina.response.HttpResponse;
+import org.apache.catalina.request.HttpRequest;
+
+public class StaticController extends AbstractController {
+
+    @Override
+    protected void doGet(HttpRequest request, HttpResponse response) {
+        response.setStatusCode(StatusCode.OK);
+        response.setBody(request.getPath());
+    }
+}
