+++ tomcat/src/main/java/org/apache/catalina/http/HttpMethod.java
@@ -1,0 +1,6 @@
+package org.apache.catalina.http;
+
+public enum HttpMethod {
+    GET,
+    POST,
+}
