+++ tomcat/src/main/java/org/apache/catalina/http/StatusCode.java
@@ -1,0 +1,25 @@
+package org.apache.catalina.http;
+
+public enum StatusCode {
+    OK(200, "OK"),
+    FOUND(302, "FOUND"),
+    UNAUTHORIZED(401, "UNAUTHORIZED"),
+    NOT_FOUND(404, "NOT FOUND"),
+    ;
+
+    private final int code;
+    private final String message;
+
+    StatusCode(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
