--- tomcat/src/main/java/org/apache/coyote/http/HttpMethod.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.http;
-
-public enum HttpMethod {
-    GET,
-    POST,
-}
