--- tomcat/src/main/java/org/apache/coyote/http/StatusCode.java
@@ -1,25 +1,0 @@
-package org.apache.coyote.http;
-
-public enum StatusCode {
-    OK(200, "OK"),
-    FOUND(302, "FOUND"),
-    UNAUTHORIZED(401, "UNAUTHORIZED"),
-    NOT_FOUND(404, "NOT FOUND"),
-    ;
-
-    private final int code;
-    private final String message;
-
-    StatusCode(int code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    public int getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
