+++ tomcat/src/test/java/org/apache/coyote/response/ResponseHeaderTest.java
@@ -2,7 +2,8 @@ package org.apache.coyote.response;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.coyote.http.HeaderName;
+import org.apache.catalina.http.HeaderName;
+import org.apache.catalina.response.ResponseHeader;
 import org.junit.jupiter.api.Test;
 
 class ResponseHeaderTest {
