+++ tomcat/src/test/java/org/apache/coyote/response/StatusLineTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote.response;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.catalina.response.StatusLine;
 import org.junit.jupiter.api.Test;
 
 class StatusLineTest {
